NPC interact Widget wasn't showing up in built project. Fixed issue
In the assest PCCCharacterSprite_Location the Interact Function WASD was unable to locate the data for the variable HUD_DialogueWASD. Oringally, this variable was established in PCCCharacterSprite_Location Event graph by casting to BP_Game_mode_locations, getting a variable and setting HUD_Dialogue as that variable. Problem was fixed by adding the same casting to the Interact Function WASD tab.